% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert.R
\name{animate_hilbert_morph}
\alias{animate_hilbert_morph}
\title{Animate a smooth transition between Hilbert curve orders}
\usage{
animate_hilbert_morph(paths)
}
\arguments{
\item{paths}{of the Hilbert curves
produced by \code{get_hilbert_paths(1:n)}}
}
\value{
a gganimate object
}
\description{
This function uses gganimate to produce an animation
of a morphing between the paths of the Hilbert curves
produced by \code{get_hilbert_paths(1:n)}
}
\examples{
get_hilbert_paths(1:3) \%>\%
  animate_hilbert_morph()
}
