% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert.R
\name{get_hilbert_path}
\alias{get_hilbert_path}
\title{Get one Hilbert path}
\usage{
get_hilbert_path(order)
}
\arguments{
\item{order, }{integer. Order of the Hilbert curve.}
}
\value{
a tibble of coordinates
}
\description{
Gets the Hilbert path for one particular order
}
