% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert.R
\name{get_hilbert_paths}
\alias{get_hilbert_paths}
\title{Get Hilbert Paths}
\usage{
get_hilbert_paths(orders)
}
\arguments{
\item{orders, }{integer vector. Orders of the Hilbert curves. E.g. \code{1:3}}
}
\value{
a tibble of coordinates
}
\description{
Get the Hilbert paths for one multiple orders
}
\examples{
get_hilbert_paths(1:2)
}
