% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbeRt-package.R
\docType{package}
\name{hilbeRt-package}
\alias{hilbeRt}
\alias{hilbeRt-package}
\title{hilbeRt: Qui pacotão da porra!}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scharth/hilbeRt}
  \item Report bugs at \url{https://github.com/scharth/hilbeRt/issues}
}

}
\keyword{internal}
