% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{coords_to_position}
\alias{coords_to_position}
\alias{coords_to_position.data.frame}
\alias{coords_to_position.matrix}
\alias{coords_to_position.numeric}
\alias{coords_to_position.integer}
\alias{coords_to_position.double}
\alias{position_to_coords}
\alias{position_to_coords.data.frame}
\alias{position_to_coords.matrix}
\alias{position_to_coords.numeric}
\alias{position_to_coords.integer}
\title{Convert Coordinates to Grid Positions}
\usage{
coords_to_position(x, ..., extent = NULL, n = 10L)

\method{coords_to_position}{data.frame}(
  x,
  ...,
  coords = c(1, 2),
  extent = NULL,
  n = 10L,
  attach = TRUE
)

\method{coords_to_position}{matrix}(
  x,
  ...,
  coords = c(1, 2),
  extent = NULL,
  n = 10L,
  attach = TRUE
)

\method{coords_to_position}{numeric}(x, y, ..., extent = NULL, n = 10L)

\method{coords_to_position}{integer}(x, y, ..., extent = NULL, n = 10L)

\method{coords_to_position}{double}(x, y, ..., extent = NULL, n = 10L)

position_to_coords(x, ..., extent = NULL, n = 10L)

\method{position_to_coords}{data.frame}(
  x,
  ...,
  coords = c(1, 2),
  extent = NULL,
  n = 10L,
  attach = TRUE
)

\method{position_to_coords}{matrix}(
  x,
  ...,
  coords = c(1, 2),
  extent = NULL,
  n = 10L,
  attach = TRUE
)

\method{position_to_coords}{numeric}(x, y, ..., extent = NULL, n = 10L)

\method{position_to_coords}{integer}(x, y, ..., extent = NULL, n = 10L)
}
\arguments{
\item{x}{One of: Integer vector, \code{data.frame}, or \code{matrix}.
If a numeric vector, then it corresponds to Row positions.}

\item{...}{Unused.}

\item{extent}{Named vector with names \code{xmax}, \code{xmin}, \code{ymax}, \code{ymin}.
Corresponds to the bounding box of the given coordinates.
If \code{extent} is \code{NULL}, then the function will throw an exception.}

\item{n}{Exponent to the dimensions of the underlying grid. The Hilbert
Curve indices are based on a \verb{2^n x 2^n} grid. This number
must be less than 15 due to the 32-bit implementation of R.}

\item{coords}{Column names or indices of a \code{data.frame}/\code{matrix} that
contain the positions.}

\item{attach}{If \code{TRUE}, adds the coordinates as new columns to the given
\code{data.frame}/\code{matrix}. This will \emph{replace} the position columns.}

\item{y}{Integer vector corresponding to Column positions.}
}
\description{
Convert Coordinates to Grid Positions

Convert Grid Positions to Coordinates
}
