% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index.data.frame}
\alias{index.matrix}
\alias{index.double}
\alias{index.integer}
\alias{index.numeric}
\title{Index positions to a Hilbert Curve}
\usage{
index(x, ..., n = 10L)

\method{index}{data.frame}(x, ..., coords = c(1, 2), n = 10L, attach = TRUE)

\method{index}{matrix}(x, ..., coords = c(1, 2), n = 10L, attach = TRUE)

\method{index}{double}(x, y, ..., n = 10L)

\method{index}{integer}(x, y, ..., n = 10L)

\method{index}{numeric}(x, y, ..., n = 10L)
}
\arguments{
\item{x}{One of: Numeric vector, \code{data.frame}, or \code{matrix}.
If a numeric vector, then it corresponds to the rows
of a position.}

\item{...}{Unused.}

\item{n}{Exponent to the dimensions of the underlying grid. The Hilbert
Curve indices are based on a \verb{2^n x 2^n} grid. This number
must be less than 15 due to the 32-bit implementation of R.}

\item{coords}{Column names or indices of a \code{data.frame}/\code{matrix} that
contain the position coordinates.}

\item{attach}{If \code{TRUE}, adds the indices as a new column to the given
\code{data.frame}/\code{matrix}. If \code{x} is a \code{data.frame}, then the
column is named \code{h}; otherwise, it is an unnamed column at
the end of the matrix.}

\item{y}{Numeric vector.
Corresponds to the columns of a position.}
}
\description{
Index positions to a Hilbert Curve
}
\seealso{
position
}
