% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position}
\alias{position}
\alias{position.data.frame}
\alias{position.matrix}
\alias{position.numeric}
\alias{position.integer}
\title{Get index positions from a Hilbert Curve}
\usage{
position(h, ..., n = 10L)

\method{position}{data.frame}(h, ..., idx = 1, n = 10L, attach = TRUE)

\method{position}{matrix}(h, ..., idx = 1, n = 10L, attach = TRUE)

\method{position}{numeric}(h, ..., n = 10L)

\method{position}{integer}(h, ..., n = 10L)
}
\arguments{
\item{h}{One of: Integer vector, \code{data.frame}, or \code{matrix}.}

\item{...}{Unused.}

\item{n}{Exponent to the dimensions of the underlying grid. The Hilbert
Curve indices are based on a \verb{2^n x 2^n} grid. This number
must be less than 15 due to the 32-bit implementation of R. This \emph{must}
be the same as the \code{n} used in \code{index}.}

\item{idx}{Column name or index containing the Hilbert Curve indices.}

\item{attach}{If \code{TRUE}, adds the position as new columns to the given
\code{data.frame}/\code{matrix}. If \code{h} is a \code{data.frame}, then the
columns are named \code{x} and \code{y}; otherwise, it is two unnamed columns at
the end of the matrix.}
}
\description{
Get index positions from a Hilbert Curve
}
\seealso{
index
}
